requirements.txt
PyPDF2==1.26.0
pillow==9.4.0
opencv-python==4.7.0.72
PyMuPDF==1.21.1
numpy==1.24.2